requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
greenlet==2.0.1
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
